refactor: consolidate auth
