+++ src/auth.ts
@@ -4,6 +4,7 @@ import { prisma } from "@/lib/db";
 import type { JWT } from "next-auth/jwt";
 import type { Session } from "next-auth";
 import Credentials from "next-auth/providers/credentials";
+import { randomBytes } from "crypto";
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
   adapter: PrismaAdapter(prisma),
@@ -64,3 +65,57 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
     },
   },
 });
+
+export async function createUser(email: string, name?: string) {
+  return await prisma.user.create({
+    data: {
+      email,
+      name,
+      reportsEnabled: true,
+      reportsFrequency: "DAILY",
+    },
+  });
+}
+
+export async function findUserByEmail(email: string) {
+  return await prisma.user.findUnique({
+    where: { email },
+  });
+}
+
+export async function createVerificationToken(email: string) {
+  const token = randomBytes(32).toString("hex");
+  const expires = new Date(Date.now() + 24 * 60 * 60 * 1000); // 24 hours
+
+  return await prisma.verificationToken.create({
+    data: {
+      identifier: email,
+      token,
+      expires,
+    },
+  });
+}
+
+export async function verifyToken(token: string) {
+  const verificationToken = await prisma.verificationToken.findUnique({
+    where: { token },
+  });
+
+  if (!verificationToken || verificationToken.expires < new Date()) {
+    return null;
+  }
+
+  const user = await prisma.user.findUnique({
+    where: { email: verificationToken.identifier },
+  });
+
+  if (!user) {
+    return null;
+  }
+
+  await prisma.verificationToken.delete({
+    where: { token },
+  });
+
+  return { user, token: verificationToken };
+}
