--- src/lib/auth-utils.ts
@@ -1,26 +1,0 @@
-import { auth } from "@/auth";
-
-export async function getCurrentUser() {
-  const session = await auth();
-  return session?.user || null;
-}
-
-export async function requireAuth() {
-  const user = await getCurrentUser();
-  if (!user) {
-    throw new Error("Unauthorized");
-  }
-  return user;
-}
-
-export async function requireAdmin() {
-  const user = await requireAuth();
-  if (!user.isAdmin) {
-    throw new Error("Admin access required");
-  }
-  return user;
-}
-
-export async function getSession() {
-  return await auth();
-}
